\name{mathtables-package}
\alias{mathtables-package}
\alias{mathtables}
\docType{package}
\title{
\packageTitle{mathtables}
}
\description{
\packageDescription{mathtables}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mathtables}
\packageIndices{mathtables}
Use this package to make tables for a child to learn arithmetic, particularly adding and multiplying.
}
\author{
\packageAuthor{mathtables}

Maintainer: \packageMaintainer{mathtables}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
